+++ payment-service/src/main/java/payments/web/rest/InvoiceResource.java
@@ -2,6 +2,7 @@ package payments.web.rest;
 
 import payments.domain.Invoice;
 import payments.repository.InvoiceRepository;
+import payments.security.SecurityUtils;
 import payments.web.rest.errors.BadRequestAlertException;
 
 import io.github.jhipster.web.util.HeaderUtil;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -98,7 +100,17 @@ public class InvoiceResource {
                 .collect(Collectors.toList());
         }
         log.debug("REST request to get all Invoices");
-        return invoiceRepository.findAll();
+        List<Invoice> resultingInvoices = new ArrayList<>();
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+
+        for(Invoice in: invoiceRepository.findAll()) {
+            if(userLogin.isPresent() && userLogin.get().equals(in.getPassengerId())) {
+                resultingInvoices.add(in);
+            }
+        }
+
+        return resultingInvoices;
     }
 
     /**
@@ -111,7 +123,15 @@ public class InvoiceResource {
     public ResponseEntity<Invoice> getInvoice(@PathVariable Long id) {
         log.debug("REST request to get Invoice : {}", id);
         Optional<Invoice> invoice = invoiceRepository.findById(id);
-        return ResponseUtil.wrapOrNotFound(invoice);
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+        if(invoice.isPresent() &&
+            userLogin.isPresent() &&
+            userLogin.get().equals(invoice.get().getPassengerId())) {
+            return ResponseUtil.wrapOrNotFound(invoice);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     /**
@@ -123,8 +143,17 @@ public class InvoiceResource {
     @DeleteMapping("/invoices/{id}")
     public ResponseEntity<Void> deleteInvoice(@PathVariable Long id) {
         log.debug("REST request to delete Invoice : {}", id);
+        Optional<Invoice> invoice = invoiceRepository.findById(id);
 
-        invoiceRepository.deleteById(id);
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+        if(invoice.isPresent() &&
+            userLogin.isPresent() &&
+            userLogin.get().equals(invoice.get().getPassengerId())) {
+            invoiceRepository.deleteById(id);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, false, ENTITY_NAME, id.toString())).build();
     }
 }
